Replaced old player with new player. Plus other small fixes
Got rid of old player, and added new player, added playerinputhandler to scene. Moved camera from old player to root of hierarchy and renamed to Main Camera. Added FP Camera which is locked to head of player (this is what main camera goes to). Changed LockedDoorWithAudio to work with InteractRaycast. Fixed visual glitch with player by adding the OnValidate function to FPController.