monaroe movement and jumpscare
made him faster and also jumpscare but wanted to add sound and shake but right now its not compiling the updated code in vs so imma just do it tomorrow before class. but jumpscare is there. also if it does freeze if u turn away and leave it will wander again or if u get close it will get u